+++ tidy3d/components/heat/data/monitor_data.py
@@ -1,17 +1,19 @@
 """Monitor level data, store the DataArrays associated with a single heat monitor."""
 from __future__ import annotations
-from typing import Union, Tuple
+from typing import Union, Tuple, Optional
 
 from abc import ABC
 
-import numpy as np
 import pydantic.v1 as pd
 
 from ..monitor import TemperatureMonitor, HeatMonitorType
 from ...base_sim.data.monitor_data import AbstractMonitorData
 from ...data.data_array import SpatialDataArray
-from ...types import ScalarSymmetry, Coordinate
+from ...data.dataset import TriangularGridDataset, TetrahedralGridDataset
+from ...types import ScalarSymmetry, Coordinate, annotate_type
 from ....constants import KELVIN
+
+from ....log import log
 
 
 class HeatMonitorData(AbstractMonitorData, ABC):
@@ -47,6 +49,7 @@ class TemperatureData(HeatMonitorData):
     Example
     -------
     >>> from tidy3d import TemperatureMonitor, SpatialDataArray
+    >>> import numpy as np
     >>> temp_data = SpatialDataArray(
     ...     np.ones((2, 3, 4)), coords={"x": [0, 1], "y": [0, 1, 2], "z": [0, 1, 2, 3]}
     ... )
@@ -61,56 +64,49 @@ class TemperatureData(HeatMonitorData):
         ..., title="Monitor", description="Temperature monitor associated with the data."
     )
 
-    temperature: SpatialDataArray = pd.Field(
-        ..., title="Temperature", description="Spatial temperature field.", units=KELVIN
+    temperature: Optional[
+        Union[SpatialDataArray, annotate_type(Union[TriangularGridDataset, TetrahedralGridDataset])]
+    ] = pd.Field(
+        ...,
+        title="Temperature",
+        description="Spatial temperature field.",
+        units=KELVIN,
     )
+
+    @pd.validator("temperature", always=True)
+    def warn_no_data(cls, val, values):
+        """Warn if no data provided."""
+
+        mnt = values.get("monitor")
+
+        if val is None:
+            log.warning(
+                f"No data is available for monitor '{mnt.name}'. This is typically caused by "
+                "monitor not intersecting any solid medium."
+            )
+
+        return val
 
     @property
     def symmetry_expanded_copy(self) -> TemperatureData:
         """Return copy of self with symmetry applied."""
 
+        if self.temperature is None:
+            return self.updated_copy(symmetry=(0, 0, 0))
+
         if all(sym == 0 for sym in self.symmetry):
             return self.copy()
 
-        coords = list(self.temperature.coords.values())
-        data = np.array(self.temperature.data)
+        new_temp = self.temperature
 
         for dim in range(3):
-            if self.symmetry[dim] == 1:
+            # do not expand monitor with zero size along symmetry direction
+            # this is done because 2d unstructured data does not support this
+            if self.symmetry[dim] == 1 and self.monitor.size[dim] != 0:
 
-                sym_center = self.symmetry_center[dim]
+                new_temp = new_temp.reflect(axis=dim, center=self.symmetry_center[dim])
 
-                if sym_center == coords[dim].data[0]:
-                    num_duplicates = 1
-                else:
-                    num_duplicates = 0
-
-                shape = np.array(np.shape(data))
-                old_len = shape[dim]
-                shape[dim] = 2 * old_len - num_duplicates
-
-                ind_left = [slice(shape[0]), slice(shape[1]), slice(shape[2])]
-                ind_right = [slice(shape[0]), slice(shape[1]), slice(shape[2])]
-
-                ind_left[dim] = slice(old_len - 1, None, -1)
-                ind_right[dim] = slice(old_len - num_duplicates, None)
-
-                new_data = np.zeros(shape)
-
-                new_data[ind_left[0], ind_left[1], ind_left[2]] = data
-                new_data[ind_right[0], ind_right[1], ind_right[2]] = data
-
-                new_coords = np.zeros(shape[dim])
-                new_coords[old_len - num_duplicates :] = coords[dim]
-                new_coords[old_len - 1 :: -1] = 2 * sym_center - coords[dim]
-
-                coords[dim] = new_coords
-                data = new_data
-
-        coords_dict = dict(zip("xyz", coords))
-        new_temperature = SpatialDataArray(data, coords=coords_dict)
-
-        return self.updated_copy(temperature=new_temperature, symmetry=(0, 0, 0))
+        return self.updated_copy(temperature=new_temp, symmetry=(0, 0, 0))
 
 
 HeatMonitorDataType = Union[TemperatureData]