+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added support for two-photon absorption via `TwoPhotonAbsorption` class. Added `KerrNonlinearity` that implements Kerr effect without third-harmonic generation.
 - Can create `PoleResidue` from LO-TO form via `PoleResidue.from_lo_to`.
+- Added `TriangularGridDataset` and `TehrahedralGridDataset` for storing and manipulating unstructured data.
 - Support for an anisotropic medium containing PEC components.
 - `SimulationData.mnt_data_from_file()` method to load only a single monitor data object from a simulation data `.hdf5` file.
 - `_hash_self` to base model, uses `hashlib` to hash a Tidy3D component the same way every session.